Add visualization regeneration feature and pre-push repo sweep
New feature:
- POST /api/generate/visualization: rate-limited, CSRF-protected endpoint
  with multimodal screenshot support and Layer 0/1 repair pipeline
- RegenerateVizModal: user feedback + screenshot upload component
- vizRegeneration prompt: per-type spec constraints for all 8 viz types
- parseLessonContent: normalizes schema-drifted lesson JSON with fallbacks
- VisualizationSectionRenderer: regenerate button + loading overlay
- i18n: viz modal keys added to en + pl locale files
- E2E tests: 4 Playwright scenarios using mock model

Cleanup:
- Gitignore 50+ one-off gen-*.mjs / assemble-*.mjs batch scripts
- Gitignore A/B test analysis artifacts
- Stage reusable utility scripts (import-gallery-local, seed-gallery,
  import-course-library) and generate-course Claude skill
- ESLint: exclude gitignored scripts; disable set-state-in-effect rule
  (intentional initialization pattern in useMobile, useLocalScratchpad, etc.)
- tsconfig: exclude ephemeral ab-test scripts from TypeScript checking

Bug fixes (pre-existing):
- notebook/route.ts: orderIndex schema incorrectly enforced .int(); Note
  model uses Float, causing all insert-custom-page calls to silently fail
- tests/e2e/dashboard: courses API returns {courses:[]} not [], causing
  TypeError in cleanup loop
- tests/integration/gallery: subject filter param renamed subjects→subject
  in earlier PR but test never updated

Docs:
- CLAUDE.md / AGENTS.md: document parseLessonContent, vizValidation,
  vizRegeneration prompt, RegenerateVizModal, and viz API endpoint

+++ src/app/api/courses/[courseId]/notebook/route.ts
@@ -6,7 +6,7 @@ import { parseBody } from "@/lib/api-validation";
 
 const createNotebookPageSchema = z.object({
   title: z.string().max(200).optional(),
-  orderIndex: z.number().int().min(0).max(1000).optional(),
+  orderIndex: z.number().min(0).max(1000).optional(),
 });
 
 interface NotebookPage {
